+++ NEWS.md
@@ -4,6 +4,10 @@
 
 * Add functionality to retrieve sub-entities from the _Central Coordinating Register for Legal Entities_ to `get_entity()` (#47)
 
+## New features
+
+* Add new function `get_schools()` (#33)
+
 # noAPI 0.1.0 (March 29 2023)
 
 This is the first release of noAPI, introducing the following functions: